pyproject.toml
[project]
name = "dedupe-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.4",
    "faker>=40.4.0",
    "matplotlib>=3.10.8",
    "mysql-connector-python>=9.6.0",
    "pandas>=3.0.0",
    "polars>=1.38.1",
    "pyarrow>=23.0.0",
    "pymysql>=1.1.2",
    "pyyaml>=6.0.3",
    "seaborn>=0.13.2",
    "splink>=4.0.13",
    "sqlalchemy>=2.0.46",
    "typer>=0.23.0",
    "unidecode>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[project.scripts]
dedupe = "dedupe_gemini:app"
